Split authentication and authorization into separate files
